+++ lib/ts/recipe/emailpassword/index.ts
@@ -129,6 +129,8 @@ export default class RecipeWrapper {
      *
      * @param formFields List of fields to send to the API exposed by the backend SDK (Refer to the {@link https://supertokens.com/docs/fdi API spec} to know more). Note that the form fields must match the ones configured in the backend SDKs
      *
+     * @param shouldTryLinkingWithSessionUser (OPTIONAL) Whether the backend should try to link the user to the session user
+     *
      * @param userContext (OPTIONAL) Refer to {@link https://supertokens.com/docs/emailpassword/advanced-customizations/user-context the documentation}
      *
      * @param options (OPTIONAL) Use this to configure additional properties (for example pre api hooks)
@@ -146,6 +148,7 @@ export default class RecipeWrapper {
             id: string;
             value: string;
         }[];
+        shouldTryLinkingWithSessionUser?: boolean;
         options?: RecipeFunctionOptions;
         userContext?: any;
     }): Promise<
@@ -170,6 +173,7 @@ export default class RecipeWrapper {
     > {
         return Recipe.getInstanceOrThrow().recipeImplementation.signUp({
             ...input,
+            shouldTryLinkingWithSessionUser: input?.shouldTryLinkingWithSessionUser,
             userContext: getNormalisedUserContext(input.userContext),
         });
     }
@@ -178,6 +182,8 @@ export default class RecipeWrapper {
      * Sign in a user with email and password
      *
      * @param formFields List of fields to send to the API exposed by the backend SDK (Refer to the {@link https://supertokens.com/docs/fdi API spec} to know more). Note that the form fields must match the ones configured in the backend SDKs
+     *
+     * @param shouldTryLinkingWithSessionUser (OPTIONAL) Whether the backend should try to link the user to the session user
      *
      * @param userContext (OPTIONAL) Refer to {@link https://supertokens.com/docs/emailpassword/advanced-customizations/user-context the documentation}
      *
@@ -198,6 +204,7 @@ export default class RecipeWrapper {
             id: string;
             value: string;
         }[];
+        shouldTryLinkingWithSessionUser?: boolean;
         options?: RecipeFunctionOptions;
         userContext?: any;
     }): Promise<
@@ -226,6 +233,7 @@ export default class RecipeWrapper {
     > {
         return Recipe.getInstanceOrThrow().recipeImplementation.signIn({
             ...input,
+            shouldTryLinkingWithSessionUser: input?.shouldTryLinkingWithSessionUser,
             userContext: getNormalisedUserContext(input.userContext),
         });
     }