+++ lib/ts/recipe/emailpassword/types.ts
@@ -150,6 +150,8 @@ export type RecipeInterface = {
      *
      * @param formFields List of fields to send to the API exposed by the backend SDK (Refer to the {@link https://supertokens.com/docs/fdi API spec} to know more). Note that the form fields must match the ones configured in the backend SDKs
      *
+     * @param shouldTryLinkingWithSessionUser Whether the backend should try to link the user to the session user
+     *
      * @param userContext Refer to {@link https://supertokens.com/docs/emailpassword/advanced-customizations/user-context the documentation}
      *
      * @param options Use this to configure additional properties (for example pre api hooks)
@@ -165,6 +167,7 @@ export type RecipeInterface = {
             id: string;
             value: string;
         }[];
+        shouldTryLinkingWithSessionUser: boolean | undefined;
         options?: RecipeFunctionOptions;
         userContext: any;
     }) => Promise<
@@ -193,6 +196,8 @@ export type RecipeInterface = {
      *
      * @param formFields List of fields to send to the API exposed by the backend SDK (Refer to the {@link https://supertokens.com/docs/fdi API spec} to know more). Note that the form fields must match the ones configured in the backend SDKs
      *
+     * @param shouldTryLinkingWithSessionUser Whether the backend should try to link the user to the session user
+     *
      * @param userContext Refer to {@link https://supertokens.com/docs/emailpassword/advanced-customizations/user-context the documentation}
      *
      * @param options Use this to configure additional properties (for example pre api hooks)
@@ -210,6 +215,7 @@ export type RecipeInterface = {
             id: string;
             value: string;
         }[];
+        shouldTryLinkingWithSessionUser: boolean | undefined;
         options?: RecipeFunctionOptions;
         userContext: any;
     }) => Promise<