+++ lib/ts/recipe/oauth2provider/types.ts
@@ -1,0 +1,87 @@
+/* Copyright (c) 2024, VRAI Labs and/or its affiliates. All rights reserved.
+ *
+ * This software is licensed under the Apache License, Version 2.0 (the
+ * "License") as published by the Apache Software Foundation.
+ *
+ * You may not use this file except in compliance with the License. You may
+ * obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+import {
+    NormalisedInputType as AuthRecipeNormalisedInputType,
+    InputType as AuthRecipeInputType,
+} from "../authRecipe/types";
+import {
+    RecipePostAPIHookContext,
+    RecipePreAPIHookContext,
+    UserInput as RecipeModuleUserInput,
+    RecipeFunctionOptions,
+} from "../recipeModule/types";
+import OverrideableBuilder from "supertokens-js-override";
+
+export type PreAndPostAPIHookAction =
+    | "GET_LOGIN_CHALLENGE_INFO"
+    | "GET_REDIRECT_URL_TO_CONTINUE_OAUTH_FLOW"
+    | "LOG_OUT";
+
+export type PreAPIHookContext = RecipePreAPIHookContext<PreAndPostAPIHookAction>;
+export type PostAPIHookContext = RecipePostAPIHookContext<PreAndPostAPIHookAction>;
+
+export type UserInput = {
+    override?: {
+        functions?: (
+            originalImplementation: RecipeInterface,
+            builder: OverrideableBuilder<RecipeInterface>
+        ) => RecipeInterface;
+    };
+} & RecipeModuleUserInput<PreAndPostAPIHookAction>;
+
+export type InputType = AuthRecipeInputType<PreAndPostAPIHookAction> & UserInput;
+
+export type NormalisedInputType = AuthRecipeNormalisedInputType<PreAndPostAPIHookAction> & {
+    override: {
+        functions: (
+            originalImplementation: RecipeInterface,
+            builder: OverrideableBuilder<RecipeInterface>
+        ) => RecipeInterface;
+    };
+};
+
+export type RecipeInterface = {
+    getLoginChallengeInfo: (input: {
+        loginChallenge: string;
+        options?: RecipeFunctionOptions;
+        userContext: any;
+    }) => Promise<{ status: "OK"; info: LoginInfo; fetchResponse: Response }>;
+    getRedirectURLToContinueOAuthFlow: (input: {
+        loginChallenge: string;
+        options?: RecipeFunctionOptions;
+        userContext: any;
+    }) => Promise<{ status: "OK"; frontendRedirectTo: string; fetchResponse: Response }>;
+    logOut: (input: {
+        logoutChallenge: string;
+        options?: RecipeFunctionOptions;
+        userContext: any;
+    }) => Promise<{ status: "OK"; frontendRedirectTo: string; fetchResponse: Response }>;
+};
+
+export type LoginInfo = {
+    clientId: string;
+    // The name of the client.
+    clientName: string;
+    // The URI of the client's terms of service.
+    tosUri?: string;
+    // The URI of the client's privacy policy.
+    policyUri?: string;
+    // The URI of the client's logo.
+    logoUri?: string;
+    // The URI of the client.
+    clientUri?: string;
+    // The metadata associated with the client.
+    metadata?: Record<string, any> | null;
+};