+++ lib/ts/recipe/passwordless/recipeImplementation.ts
@@ -28,11 +28,7 @@ export default function getRecipeImplementation(
     const querier = new Querier(recipeImplInput.recipeId, recipeImplInput.appInfo);
 
     return {
-        createCode: async function (
-            input:
-                | { email: string; userContext: any; options?: RecipeFunctionOptions }
-                | { phoneNumber: string; userContext: any; options?: RecipeFunctionOptions }
-        ): Promise<
+        createCode: async function (input): Promise<
             | {
                   status: "OK";
                   deviceId: string;
@@ -51,12 +47,14 @@ export default function getRecipeImplementation(
             if ("email" in input) {
                 bodyObj = {
                     email: input.email,
+                    shouldTryLinkingWithSessionUser: input.shouldTryLinkingWithSessionUser,
                 };
             }
 
             if ("phoneNumber" in input) {
                 bodyObj = {
                     phoneNumber: input.phoneNumber,
+                    shouldTryLinkingWithSessionUser: input.shouldTryLinkingWithSessionUser,
                 };
             }
 
@@ -93,6 +91,7 @@ export default function getRecipeImplementation(
             tenantId: string | undefined;
             deviceId: string;
             preAuthSessionId: string;
+            shouldTryLinkingWithSessionUser: boolean | undefined;
             userContext: any;
             options?: RecipeFunctionOptions;
         }): Promise<{
@@ -102,6 +101,7 @@ export default function getRecipeImplementation(
             const bodyObj = {
                 deviceId: input.deviceId,
                 preAuthSessionId: input.preAuthSessionId,
+                shouldTryLinkingWithSessionUser: input.shouldTryLinkingWithSessionUser,
             };
 
             const { jsonBody, fetchResponse } = await querier.post<{
@@ -128,24 +128,7 @@ export default function getRecipeImplementation(
                 fetchResponse,
             };
         },
-        consumeCode: async function (
-            input:
-                | {
-                      tenantId: string | undefined;
-                      userInputCode: string;
-                      deviceId: string;
-                      preAuthSessionId: string;
-                      userContext: any;
-                      options?: RecipeFunctionOptions;
-                  }
-                | {
-                      tenantId: string | undefined;
-                      preAuthSessionId: string;
-                      linkCode: string;
-                      userContext: any;
-                      options?: RecipeFunctionOptions;
-                  }
-        ): Promise<
+        consumeCode: async function (input): Promise<
             | {
                   status: "OK";
                   createdNewRecipeUser: boolean;
@@ -167,11 +150,13 @@ export default function getRecipeImplementation(
                     userInputCode: input.userInputCode,
                     deviceId: input.deviceId,
                     preAuthSessionId: input.preAuthSessionId,
+                    shouldTryLinkingWithSessionUser: input.shouldTryLinkingWithSessionUser,
                 };
             } else {
                 bodyObj = {
                     linkCode: input.linkCode,
                     preAuthSessionId: input.preAuthSessionId,
+                    shouldTryLinkingWithSessionUser: input.shouldTryLinkingWithSessionUser,
                 };
             }
 
@@ -317,6 +302,7 @@ export default function getRecipeImplementation(
                   tenantId?: string;
                   preAuthSessionId: string;
                   flowType: PasswordlessFlowType;
+                  shouldTryLinkingWithSessionUser?: boolean;
               } & CustomLoginAttemptInfoProperties)
         > {
             const storedInfo = await WindowHandlerReference.getReferenceOrThrow().windowHandler.localStorage.getItem(