+++ lib/ts/recipe/passwordless/types.ts
@@ -67,6 +67,8 @@ export type RecipeInterface = {
      *
      * @param phoneNumber Phone number of the user
      *
+     * @param shouldTryLinkingWithSessionUser Whether the backend should try to link the user to the session user when the code is consumed
+     *
      * @param userContext Refer to {@link https://supertokens.com/docs/passwordless/advanced-customizations/user-context the documentation}
      *
      * @param options Use this to configure additional properties (for example pre api hooks)
@@ -77,8 +79,18 @@ export type RecipeInterface = {
      */
     createCode: (
         input:
-            | { email: string; userContext: any; options?: RecipeFunctionOptions }
-            | { phoneNumber: string; userContext: any; options?: RecipeFunctionOptions }
+            | {
+                  email: string;
+                  shouldTryLinkingWithSessionUser: boolean | undefined;
+                  userContext: any;
+                  options?: RecipeFunctionOptions;
+              }
+            | {
+                  phoneNumber: string;
+                  shouldTryLinkingWithSessionUser: boolean | undefined;
+                  userContext: any;
+                  options?: RecipeFunctionOptions;
+              }
     ) => Promise<
         | {
               status: "OK";
@@ -101,6 +113,8 @@ export type RecipeInterface = {
      *
      * @param preAuthSessionId The id from the response of `createCode`
      *
+     * @param shouldTryLinkingWithSessionUser Whether the backend should try to link the user to the session user when the code is consumed
+     *
      * @param userContext Refer to {@link https://supertokens.com/docs/passwordless/advanced-customizations/user-context the documentation}
      *
      * @param options Use this to configure additional properties (for example pre api hooks)
@@ -115,6 +129,7 @@ export type RecipeInterface = {
         userContext: any;
         deviceId: string;
         preAuthSessionId: string;
+        shouldTryLinkingWithSessionUser: boolean | undefined;
         tenantId: string | undefined;
         options?: RecipeFunctionOptions;
     }) => Promise<{
@@ -130,6 +145,8 @@ export type RecipeInterface = {
      * @param deviceId The device if from the reponse of `createCode`. (Not required when using `linkCode`)
      *
      * @param preAuthSessionId The id from the response of `createCode`.
+     *
+     * @param shouldTryLinkingWithSessionUser Whether the backend should try to link the user to the session user
      *
      * @param linkCode The code from the URL to use when logging the user in. Ignored if `userInputCode` is provided
      *
@@ -155,12 +172,14 @@ export type RecipeInterface = {
                   deviceId: string;
                   tenantId: string | undefined;
                   preAuthSessionId: string;
+                  shouldTryLinkingWithSessionUser: boolean | undefined;
                   userContext: any;
                   options?: RecipeFunctionOptions;
               }
             | {
                   tenantId: string | undefined;
                   preAuthSessionId: string;
+                  shouldTryLinkingWithSessionUser: boolean | undefined;
                   linkCode: string;
                   userContext: any;
                   options?: RecipeFunctionOptions;
@@ -264,6 +283,7 @@ export type RecipeInterface = {
               tenantId?: string;
               deviceId: string;
               preAuthSessionId: string;
+              shouldTryLinkingWithSessionUser?: boolean;
               flowType: PasswordlessFlowType;
           } & CustomLoginAttemptInfoProperties)
     >;
@@ -278,6 +298,7 @@ export type RecipeInterface = {
             tenantId?: string;
             deviceId: string;
             preAuthSessionId: string;
+            shouldTryLinkingWithSessionUser: boolean | undefined;
             flowType: PasswordlessFlowType;
         } & CustomStateProperties;
         userContext: any;